+++ tools/lib_alignments/annotate.py
@@ -42,8 +42,7 @@ class Annotate():
             bottom_right = (alignment["x"] + alignment["w"], alignment["y"] + alignment["h"])
             logger.trace("Drawing bounding box: (top_left: %s, bottom_right: %s, color: %s, "
                          "thickness: %s)", top_left, bottom_right, color, thickness)
-            cv2.rectangle(self.image,  # pylint: disable=no-member
-                          top_left, bottom_right, color, thickness)
+            cv2.rectangle(self.image, top_left, bottom_right, color, thickness)
 
     def draw_extract_box(self, color_id=2, thickness=1):
         """ Draw the extracted face box """
@@ -54,25 +53,24 @@ class Annotate():
             logger.trace("Drawing Extract Box: (idx: %s, roi: %s)", idx, roi)
             top_left = [point for point in roi.squeeze()[0]]
             top_left = (top_left[0], top_left[1] - 10)
-            cv2.putText(self.image,  # pylint: disable=no-member
+            cv2.putText(self.image,
                         str(idx),
                         top_left,
-                        cv2.FONT_HERSHEY_DUPLEX,  # pylint: disable=no-member
+                        cv2.FONT_HERSHEY_DUPLEX,
                         1.0,
                         color,
                         thickness)
-            cv2.polylines(self.image, [roi], True, color, thickness)  # pylint: disable=no-member
+            cv2.polylines(self.image, [roi], True, color, thickness)
 
     def draw_landmarks(self, color_id=3, radius=1):
         """ Draw the facial landmarks """
         color = self.colors[color_id]
         for alignment in self.alignments:
-            landmarks = alignment["landmarks_xy"]
+            landmarks = alignment["landmarks_xy"].astype("int32")
             logger.trace("Drawing Landmarks: (landmarks: %s, color: %s, radius: %s)",
                          landmarks, color, radius)
             for (pos_x, pos_y) in landmarks:
-                cv2.circle(self.image,  # pylint: disable=no-member
-                           (pos_x, pos_y), radius, color, -1)
+                cv2.circle(self.image, (pos_x, pos_y), radius, color, -1)
 
     def draw_landmarks_mesh(self, color_id=4, thickness=1):
         """ Draw the facial landmarks """
@@ -92,8 +90,7 @@ class Annotate():
             for key, val in facial_landmarks_idxs.items():
                 points = np.array([landmarks[val[0]:val[1]]], np.int32)
                 fill_poly = bool(key in ("right_eye", "left_eye", "mouth"))
-                cv2.polylines(self.image,  # pylint: disable=no-member
-                              points, fill_poly, color, thickness)
+                cv2.polylines(self.image, points, fill_poly, color, thickness)
 
     def draw_grey_out_faces(self, live_face):
         """ Grey out all faces except target """
@@ -104,7 +101,6 @@ class Annotate():
         for idx, roi in enumerate(self.roi):
             if idx != int(live_face):
                 logger.trace("Greying out face: (idx: %s, roi: %s)", idx, roi)
-                cv2.fillPoly(overlay, roi, (0, 0, 0))  # pylint: disable=no-member
+                cv2.fillPoly(overlay, roi, (0, 0, 0))
 
-        cv2.addWeighted(overlay,  # pylint: disable=no-member
-                        alpha, self.image, 1. - alpha, 0., self.image)
+        cv2.addWeighted(overlay, alpha, self.image, 1. - alpha, 0., self.image)
